refactor: move Python placeholder manifests into monochange_python
