+++ .changeset/move-python-placeholder-manifest.md
@@ -1,0 +1,8 @@
+---
+monochange: patch
+monochange_python: minor
+---
+
+# Move Python placeholder manifests into monochange_python
+
+Move PyPI placeholder `pyproject.toml` and module scaffold generation out of `monochange::package_publish` and into `monochange_python`.
