+++ crates/monochange_python/Cargo.toml
@@ -15,6 +15,7 @@ description = "Python ecosystem adapter for monochange — discovers uv workspac
 [dependencies]
 glob = { workspace = true, default-features = true }
 monochange_core = { workspace = true }
+monochange_publish = { workspace = true }
 semver = { workspace = true, default-features = true }
 serde = { workspace = true, default-features = true }
 toml = { workspace = true, default-features = true }
